Modifications to get murfi to run on 64-bit OSs
The header doesn't use OS indepentent types, so sizeof(long) changes between 32 abd 64-bit OSs.

Ideally, the header would be re-written to use OS independent types (ie int32_t) but that'd be alot of work to ensure everything is casted properly all over the code.  Instead, as a temporary measure, RtExternalImageInfo has been modified to read in int32_t instead of longs, then cast the resulting value to a long and store it in the proper place.