add: Dockerfile that builds many dependencies from source
The dependencies built from source are:
  - cmake (a dependency of some dependencies)
  - vxl
  - boost (filesystem and system)
  - nifticlib
  - gsl

Qt5 is installed from apt repositories, and libace-dev is installed from apt as well.

This Dockerfile also makes use of multi-stage builds. In the first stage, murfi's dependencies and murfi are compiled. In the second stage, murfi and its artifacts are copied over, essential things from compiled dependencies are copied over, and qt5 and libace are installed from apt repositories. A Docker image can be made from the first stage, if desired. For reference, the first-stage Docker image is 2.43 GB, whereas the Docker image from both stages is 297 MB (both values are uncompressed sizes).